Moved Plane to Util, changed Plane constructor to take Double3D, made necessary adjustments
