+++ test/arcade/core/util/PlaneTest.java
@@ -1,7 +1,7 @@
-package arcade.potts.env.location;
+package arcade.core.util;
 
 import org.junit.jupiter.api.Test;
-import arcade.core.util.Vector;
+import sim.util.Double3D;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class PlaneTest {
@@ -9,7 +9,7 @@ public class PlaneTest {
 
     @Test
     public void constructor_givenPointAndUnitVector_returnsCorrectPlane() {
-        Voxel point = new Voxel(1, 2, 3);
+        Double3D point = new Double3D(1, 2, 3);
         Vector normalVector = new Vector(1, 0, 0);
         Plane plane = new Plane(point, normalVector);
         assertEquals(point, plane.referencePoint);
@@ -18,7 +18,7 @@ public class PlaneTest {
 
     @Test
     public void constructor_givenPointAndIntegerVector_returnsCorrectPlane() {
-        Voxel point = new Voxel(1, 2, 3);
+        Double3D point = new Double3D(1, 2, 3);
         Vector normalVector = new Vector(2, 2, 1);
 
         Plane plane = new Plane(point, normalVector);
@@ -33,7 +33,7 @@ public class PlaneTest {
 
     @Test
     public void constructor_givenPointAndDoubleVector_returnsCorrectPlane() {
-        Voxel point = new Voxel(1, 2, 3);
+        Double3D point = new Double3D(1, 2, 3);
         Vector normalVector = new Vector(1.5, 1.5, 1);
 
         Plane plane = new Plane(point, normalVector);
@@ -76,61 +76,61 @@ public class PlaneTest {
 
     @Test
     public void equals_givenDifferentPlane_returnsFalse() {
-        Plane plane1 = new Plane(new Voxel(1, 2, 3), new Vector(4, 5, 6));
-        Plane plane2 = new Plane(new Voxel(4, 5, 6), new Vector(7, 8, 9));
+        Plane plane1 = new Plane(new Double3D(1, 2, 3), new Vector(4, 5, 6));
+        Plane plane2 = new Plane(new Double3D(4, 5, 6), new Vector(7, 8, 9));
 
         assertNotEquals(plane1, plane2);
     }
 
     @Test
     public void equals_givenSamePlane_returnsTrue() {
-        Plane plane1 = new Plane(new Voxel(1, 2, 3), new Vector(4, 5, 6));
-        Plane plane2 = new Plane(new Voxel(1, 2, 3), new Vector(4, 5, 6));
+        Plane plane1 = new Plane(new Double3D(1, 2, 3), new Vector(4, 5, 6));
+        Plane plane2 = new Plane(new Double3D(1, 2, 3), new Vector(4, 5, 6));
 
         assertEquals(plane1, plane2);
     }
 
     @Test
     public void distanceToPlane_givenPointOnPlane_returnsZero() {
-        Voxel pointOnPlane = new Voxel(0, 0, 0);
+        Double3D pointOnPlane = new Double3D(0, 0, 0);
         Vector normalVector = new Vector(1, 0, 0);
         Plane plane = new Plane(pointOnPlane, normalVector);
 
-        Voxel pointToTest = new Voxel(0, 0, 0);
+        Double3D pointToTest = new Double3D(0, 0, 0);
 
         assertEquals(0, plane.signedDistanceToPlane(pointToTest), EPSILON);
     }
 
     @Test
     public void distanceToPlane_givenPointOnNormalSideOfPlane_returnsCorrectPositiveDistance() {
-        Voxel pointOnPlane = new Voxel(0, 0, 0);
+        Double3D pointOnPlane = new Double3D(0, 0, 0);
         Vector normalVector = new Vector(1, 0, 0);
         Plane plane = new Plane(pointOnPlane, normalVector);
 
-        Voxel pointToTest = new Voxel(1, 1, 1);
+        Double3D pointToTest = new Double3D(1, 1, 1);
 
         assertEquals(1, plane.signedDistanceToPlane(pointToTest), EPSILON);
     }
 
     @Test
     public void distanceToPlane_givenPointOnOppositeSideOfPlane_returnsCorrectNegativeDistance() {
-        Voxel pointOnPlane = new Voxel(0, 0, 0);
+        Double3D pointOnPlane = new Double3D(0, 0, 0);
         Vector normalVector = new Vector(1, 0, 0);
         Plane plane = new Plane(pointOnPlane, normalVector);
 
-        Voxel pointToTest = new Voxel(-1, -1, -1);
+        Double3D pointToTest = new Double3D(-1, -1, -1);
 
         assertEquals(-1, plane.signedDistanceToPlane(pointToTest), EPSILON);
     }
 
     @Test
     public void signedDistanceToPlane_givenPointOnAxis_returnsCorrectSignedDistance() {
-        Voxel point = new Voxel(0, 0, 0);
+        Double3D point = new Double3D(0, 0, 0);
         Vector normalVector = new Vector(0, 0, 1);
         Plane plane = new Plane(point, normalVector);
 
-        Voxel posTestPoint = new Voxel(0, 0, 5);
-        Voxel negTestPoint = new Voxel(0, 0, -5);
+        Double3D posTestPoint = new Double3D(0, 0, 5);
+        Double3D negTestPoint = new Double3D(0, 0, -5);
 
         assertEquals(5.0, plane.signedDistanceToPlane(posTestPoint), EPSILON);
         assertEquals(-5.0, plane.signedDistanceToPlane(negTestPoint), EPSILON);
@@ -138,12 +138,12 @@ public class PlaneTest {
 
     @Test
     public void signedDistanceToPlane_givenPointOffAxis_returnsCorrectSignedDistance() {
-        Voxel point = new Voxel(1, 1, 2);
+        Double3D point = new Double3D(1, 1, 2);
         Vector normalVector = new Vector(1, 2, 2);
         Plane plane = new Plane(point, normalVector);
 
-        Voxel postestPoint = new Voxel(2, 2, 5);
-        Voxel negtestPoint = new Voxel(0, 0, -1);
+        Double3D postestPoint = new Double3D(2, 2, 5);
+        Double3D negtestPoint = new Double3D(0, 0, -1);
 
         assertEquals(3.0, plane.signedDistanceToPlane(postestPoint), EPSILON);
         assertEquals(-3.0, plane.signedDistanceToPlane(negtestPoint), EPSILON);
@@ -151,7 +151,7 @@ public class PlaneTest {
 
     @Test
     public void hashCode_equalObjects_returnsSameCode() {
-        Voxel point = new Voxel(1, 2, 3);
+        Double3D point = new Double3D(1, 2, 3);
         Vector normalVector = new Vector(4, 5, 6);
 
         Plane plane1 = new Plane(point, normalVector);
